percentage changes.

+++ src/main/java/com/ontarget/bean/Task.java
@@ -16,7 +16,7 @@ public class Task implements Serializable{
     private String severity;
     private double cost;
     private int numberOfWorkers;
-    private int percentageComplete;
+    private double percentageComplete;
     private String startDateText;
     private String endDateText;
 
@@ -118,11 +118,11 @@ public class Task implements Serializable{
         this.numberOfWorkers = numberOfWorkers;
     }
 
-    public int getPercentageComplete() {
+    public double getPercentageComplete() {
         return percentageComplete;
     }
 
-    public void setPercentageComplete(int percentageComplete) {
+    public void setPercentageComplete(double percentageComplete) {
         this.percentageComplete = percentageComplete;
     }
